requirements.txt
numpy~=1.26.4
opencv-python~=4.10.0.84
matplotlib
pillow~=11.0.0
scikit-image~=0.24.0
torch~=2.5.0
torchvision~=0.20.0
flask~=3.0.3
scikit-learn~=1.5.2